refactor: move all llama code to models/llama out of meta reference
