+++ llama_stack/models/llama/llama3/multimodal/encoder_utils.py
