+++ llama_stack/models/llama/llama3/multimodal/image_transform.py
