+++ llama_stack/models/llama/llama3/multimodal/model.py
