+++ llama_stack/models/llama/llama4/vision/embedding.py
