+++ llama_stack/models/llama/llama4/vision/encoder.py
